+++ README.md
@@ -11,7 +11,7 @@ A clean, simple and lightweight HTML parser built on top of
 
 - follows the [spec](#spec) closely
 - parse [elements](https://jsr.io/@fcrozatier/htmlcrunch/doc/~/element),
-  [fragments](https://jsr.io/@fcrozatier/htmlcrunch/doc/~/fragments) and whole
+  [fragments](https://jsr.io/@fcrozatier/htmlcrunch/doc/~/fragments) and complete
   [html](https://jsr.io/@fcrozatier/htmlcrunch/doc/~/html) documents
 - transform the parse tree and use
   [`isCommentNode`](https://jsr.io/@fcrozatier/htmlcrunch/doc/~/isCommentNode),
@@ -52,27 +52,27 @@ assertEquals(content, serialized);
 HtmlCrunch implements the following parts of the
 [HTML spec](https://html.spec.whatwg.org/):
 
-| spec                                                                                           | status             |
-| ---------------------------------------------------------------------------------------------- | ------------------ |
-| [document structure](https://html.spec.whatwg.org/#writing)                                    | ✅                 |
-| [modern doctype](https://html.spec.whatwg.org/#the-doctype)                                    | ✅                 |
-| [legacy doctype](https://html.spec.whatwg.org/#doctype-legacy-string)                          | 🚫 (not planned)   |
-| **[Elements](https://html.spec.whatwg.org/#elements-2)**                                       |                    |
-| self-closing void elements                                                                     | ✅                 |
-| raw text elements                                                                              | ✅                 |
-| foreign elements (MathML & SVG namespaces)                                                     | ✅                 |
-| normal elements                                                                                | ✅                 |
-| **[Attributes](https://html.spec.whatwg.org/#attributes-2)**                                   |                    |
-| Empty attribute syntax                                                                         | ✅                 |
-| Unquoted attribute value syntax                                                                | ✅                 |
-| Single-quoted attribute value syntax                                                           | ✅                 |
-| Double-quoted attribute value syntax                                                           | ✅                 |
-| **[Optional tags]((https://html.spec.whatwg.org/#syntax-tag-omission))**                       |                    |
-| end tag omission                                                                               | ✅                 |
-| start tag omission                                                                             | 🚫 (not planned)   |
+| spec                                                                                           | status            |
+| ---------------------------------------------------------------------------------------------- | ----------------- |
+| **Structure**                                                                                  |                   |
+| - [document structure](https://html.spec.whatwg.org/#writing)                                  | ✅                 |
+| - [modern doctype](https://html.spec.whatwg.org/#the-doctype)                                  | ✅                 |
+| **[Elements](https://html.spec.whatwg.org/#elements-2)**                                       |                   |
+| - self-closing void elements                                                                   | ✅                 |
+| - raw text elements                                                                            | ✅                 |
+| - foreign elements (MathML & SVG namespaces)                                                   | ✅                 |
+| - normal elements                                                                              | ✅                 |
+| **[Attributes](https://html.spec.whatwg.org/#attributes-2)**                                   |                   |
+| - Empty attribute syntax                                                                       | ✅                 |
+| - Unquoted attribute value syntax                                                              | ✅                 |
+| - Single-quoted attribute value syntax                                                         | ✅                 |
+| - Double-quoted attribute value syntax                                                         | ✅                 |
+| **[Optional tags]((https://html.spec.whatwg.org/#syntax-tag-omission))**                       |                   |
+| - end tag omission                                                                             | ✅                 |
+| - start tag omission                                                                           | 🚫 (not planned)   |
 | content model validation and [restriction](https://html.spec.whatwg.org/#element-restrictions) | ⚠️ (not supported) |
 | [text](https://html.spec.whatwg.org/#text-2)                                                   | ✅                 |
-| [CDATA sections](https://html.spec.whatwg.org/#cdata-sections)                                 | ⚠️ (not supported) |
+| [CDATA sections](https://html.spec.whatwg.org/#cdata-sections)                                 | ✅                 |
 | [comments](https://html.spec.whatwg.org/#comments)                                             | ✅                 |
 
 ## End tag omission
